requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
google-generativeai==0.8.5
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
